Handle case for test customeros domain
