+++ packages/server/customer-os-api/service/namecheap_service.go
@@ -4,8 +4,8 @@ import (
 	"encoding/xml"
 	"fmt"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/config"
-	coserrors "github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/errors"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/repository"
+	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/coserrors"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/logger"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/tracing"
 	"github.com/opentracing/opentracing-go"
