+++ packages/server/customer-os-common-module/config/global_config.go
@@ -12,10 +12,15 @@ type GlobalConfig struct {
 
 	// Customer OS
 	InternalServices InternalServices
+	ExternalServices ExternalServices
 }
 
 type InternalServices struct {
 	UserAdminApiPublicPath string `env:"USER_ADMIN_API_PUBLIC_PATH,required" envDefault:"http://localhost:4001"`
 	EnrichmentApiConfig    EnrichmentAPIConfig
 	AiApiConfig            AiAPIConfig
+}
+
+type ExternalServices struct {
+	OpenSRSConfig OpenSRSConfig
 }
