+++ packages/server/customer-os-common-module/coserrors/errors.go
@@ -5,11 +5,13 @@ import (
 )
 
 var (
-	ErrAccessDenied      = errors.New("Access denied")
-	ErrInvalidEntityType = errors.New("Invalid entity type")
-	ErrNotSupported      = errors.New("Not supported")
-	ErrConnectionTimeout = errors.New("Connection timeout")
+	ErrAccessDenied        = errors.New("Access denied")
+	ErrInvalidEntityType   = errors.New("Invalid entity type")
+	ErrNotSupported        = errors.New("Not supported")
+	ErrConnectionTimeout   = errors.New("Connection timeout")
+	ErrOperationNotAllowed = errors.New("Operation not allowed")
 
+	// domain errors
 	ErrDomainUnavailable         = errors.New("domain unavailable")
 	ErrDomainPremium             = errors.New("domain is premium")
 	ErrDomainPriceExceeded       = errors.New("domain price exceeds the maximum allowed price")
@@ -17,5 +19,6 @@ var (
 	ErrDomainConfigurationFailed = errors.New("domain configuration failed")
 	ErrDomainNotFound            = errors.New("domain not found")
 
+	// mailbox errors
 	ErrMailboxExists = errors.New("mailbox already exists")
 )
