--- packages/server/customer-os-common-module/errors/common_errors.go
@@ -1,7 +1,0 @@
-package commonerrors
-
-import "github.com/pkg/errors"
-
-var (
-	ErrOperationNotAllowed = errors.New("Operation not allowed")
-)
